pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "asyncio>=3.4.3",
    "discord-py>=2.5.2",
    "pynacl>=1.5.0",
    "python-dotenv>=1.1.0",
    "wavelink>=3.4.1",
    "yt-dlp>=2025.3.31",
]
